--- app/src/main/java/com/sopt/now/compose/ext/NavigationExt.kt
@@ -1,11 +1,0 @@
-package com.sopt.now.compose.ext
-
-import androidx.navigation.NavHostController
-import androidx.navigation.NavOptionsBuilder
-
-
-fun NavOptionsBuilder.removePreviousBackStacks(navController: NavHostController, screenInclusive: Boolean = true) {
-    popUpTo(navController.currentBackStackEntry?.destination?.route!!)
-    { inclusive = screenInclusive }
-}
-
