--- app/src/main/java/com/sopt/now/compose/ui/screens/login/LoginUiState.kt
@@ -1,8 +1,0 @@
-package com.sopt.now.compose.ui.screens.login
-
-data class LoginUiState(
-    val id: String = "",
-    val password: String = "",
-    val isSuccess: Boolean = false,
-    val message:String = ""
-)
