--- app/src/main/java/com/sopt/now/compose/ui/screens/signup/SignUpState.kt
@@ -1,10 +1,0 @@
-package com.sopt.now.compose.ui.screens.signup
-
-data class SignUpState(
-    val isSuccess: Boolean = false,
-    val message: String = "",
-    val authenticationId: String = "",
-    val password: String = "",
-    val nickName: String = "",
-    val phone: String = "",
-)
