+++ presentation/src/main/java/com/sopt/now/compose/presentation/composables/ButtonComposable.kt
@@ -1,0 +1,30 @@
+package com.sopt.now.compose.presentation.composables
+
+import androidx.annotation.StringRes
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ButtonComposable(
+    @StringRes
+    text: Int,
+    color: Color = Color.Unspecified,
+    onClick: () -> Unit = {},
+) {
+    Button(
+        modifier = Modifier.fillMaxWidth(),
+        onClick = onClick,
+        colors = ButtonDefaults.buttonColors(containerColor = color),
+        shape = RoundedCornerShape(8.dp)
+    ) {
+        Text(text = stringResource(id = text))
+    }
+}
