+++ presentation/src/main/java/com/sopt/now/compose/presentation/composables/TextFieldWithTitleComposable.kt
@@ -1,0 +1,46 @@
+package com.sopt.now.compose.presentation.composables
+
+import androidx.annotation.StringRes
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.sp
+
+@Composable
+fun TextFieldWithTitleComposable(
+    modifier: Modifier = Modifier,
+    @StringRes
+    title: Int,
+    @StringRes
+    label: Int,
+    titleFontSize: TextUnit = 15.sp,
+    textFieldText: String = "",
+    onValueChange: (String) -> Unit = {},
+    keyboardType: KeyboardType = KeyboardType.Ascii,
+    imeAction: ImeAction = ImeAction.Next,
+    visualTransformation: VisualTransformation = VisualTransformation.None
+) {
+    Column(
+        modifier = modifier
+    ) {
+        Text(text = stringResource(id = title), fontSize = titleFontSize)
+        TextField(
+            value = textFieldText,
+            onValueChange = onValueChange,
+            modifier = Modifier.fillMaxWidth(),
+            label = { Text(text = stringResource(id = label)) },
+            singleLine = true,
+            visualTransformation = visualTransformation,
+            keyboardOptions = KeyboardOptions(keyboardType = keyboardType, imeAction = imeAction)
+        )
+    }
+}
