+++ presentation/src/main/java/com/sopt/now/compose/presentation/navigation/NavigationDestination.kt
@@ -1,0 +1,12 @@
+package com.sopt.now.compose.presentation.navigation
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+interface NavigationDestination {
+    /**
+     * Unique name to define the path for a screen composable
+     */
+    val route: String
+    val titleRes: Int
+    val iconVector: ImageVector?
+}
