+++ presentation/src/main/java/com/sopt/now/compose/presentation/screens/login/LoginUiState.kt
@@ -1,0 +1,8 @@
+package com.sopt.now.compose.presentation.ui.screens.login
+
+data class LoginUiState(
+    val id: String = "",
+    val password: String = "",
+    val isSuccess: Boolean = false,
+    val message:String = ""
+)
