+++ presentation/src/main/java/com/sopt/now/compose/presentation/screens/signup/SignUpState.kt
@@ -1,0 +1,10 @@
+package com.sopt.now.compose.presentation.screens.signup
+
+data class SignUpState(
+    val isSuccess: Boolean = false,
+    val message: String = "",
+    val authenticationId: String = "",
+    val password: String = "",
+    val nickName: String = "",
+    val phone: String = "",
+)
